+++ lib/src/utils/media_upload/providers/video_upload_provider.dart
@@ -5,6 +5,7 @@ import 'package:core/core.dart';
 import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_trc/src/channel/native_communication.dart';
+import 'package:flutter_trc/src/utils/connectivity_util.dart';
 import 'package:flutter_trc/src/utils/media_upload/resource/media_content_type.dart';
 import 'package:path/path.dart' as path;
 import 'package:path_provider/path_provider.dart';
@@ -25,35 +26,38 @@ class VideoUploadProvider extends CshChangeNotifier {
   File? get videoThumbnailFile =>
       (!Validator.isNullOrEmpty(_videoThumbnailImagePath)) ? File(_videoThumbnailImagePath!) : null;
 
-  Future<(String, String?)> uploadVideo(File file) async {
+  Future<(String, String?)> uploadVideo(File file, {bool isCompressed = true}) async {
+    var isConnected = await ConnectivityUtil.checkConnectivity();
+    if (!isConnected) {
+      return Future.error("No Internet connection");
+    }
     isDataLoading = true;
     notifyListeners();
     var completer = Completer<(String, String?)>();
 
-    NativeCommunication.compressVideo(file.path, includeAudio: false).then((value) {
-      if (value?.file != null) {
-        file = value!.file!;
+    if (isCompressed) {
+      var mediaInfo = await NativeCommunication.compressVideo(file.path, includeAudio: false);
+      if (mediaInfo?.file != null) {
+        file = mediaInfo!.file!;
+      }
+    }
+
+    String fileName = path.basename(file.path);
+    MediaUploadUtil().uploadMediaWithType(mediaFile: file, fileName: fileName, contentType: MediaContentType.mp4).then(
+        (value) async {
+      if (value.isNotEmpty) {
+        videoS3Url = value;
+        _videoThumbnailImagePath = await _getVideoThumbnail(file.path);
+        completer.complete((value, _videoThumbnailImagePath));
+      } else {
+        completer.completeError("Something went wrong");
       }
     }, onError: (error) {
       Logger.debug('mydebug-----VideoUploadProvider.uploadVideo', [error]);
+      completer.completeError(error);
     }).whenComplete(() {
-      String fileName = path.basename(file.path);
-      MediaUploadUtil()
-          .uploadMediaWithType(mediaFile: file, fileName: fileName, contentType: MediaContentType.mp4)
-          .then((value) async {
-        if (value.isNotEmpty) {
-          videoS3Url = value;
-          _videoThumbnailImagePath = await _getVideoThumbnail(file.path);
-          completer.complete((value, _videoThumbnailImagePath));
-        } else {
-          completer.completeError("Something went wrong");
-        }
-      }, onError: (error) {
-        completer.completeError(error);
-      }).whenComplete(() {
-        isDataLoading = false;
-        notifyListeners();
-      });
+      isDataLoading = false;
+      notifyListeners();
     });
     return completer.future;
   }